[Feature] [+] = Ennemy can shoot, player decrease health and can die File Created : config/entities/11.json, config/entities/15.json, ressources/shoot-spritesheet-ennemy.png, src/ecs/Components/ShootIntervalComponent.cpp, src/ecs/Components/ShootIntervalComponent.hh File Modified : config/entities/13.json, config/entities/14.json, config/entities/4.json, src/client/Game.cpp, src/client/Game.hh, src/ecs/CMakeLists.txt, src/ecs/Components/ActionComponent.cpp, src/ecs/Components/ActionComponent.hh, src/ecs/Components/EntityTypeComponent.hh, src/ecs/Systems/HandleColisionSystem.cpp, src/ecs/Systems/HandleEntitySpawnSystem.cpp, src/ecs/Systems/HandleEntitySpawnSystem.hh, src/ecs/Systems/HandleShootSystem.cpp, src/ecs/Systems/HandleShootSystem.hpp, src/server/server/Server.cpp File Deleted : config/entities/12.json

+++ src/ecs/Systems/HandleShootSystem.cpp
@@ -7,8 +7,8 @@
 
 #include "HandleShootSystem.hpp"
 
-RType::HandleShootSystem::HandleShootSystem(std::function<std::shared_ptr<Entity>()> addEntity, std::function<void(std::shared_ptr<Entity>)> deleteEntity, std::function<void(const std::basic_string<unsigned char> &message)> sendMessageToServer):
-    ASystem(S_SHOOT, addEntity, deleteEntity), _sendMessageToServer(sendMessageToServer)
+RType::HandleShootSystem::HandleShootSystem(std::function<std::shared_ptr<Entity>()> addEntity, std::function<void(std::shared_ptr<Entity>)> deleteEntity, std::function<void(const std::basic_string<unsigned char> &message)> sendMessageToServer, std::function<void(const std::basic_string<unsigned char> &message)> sendMessageToAllClient):
+    ASystem(S_SHOOT, addEntity, deleteEntity), _sendMessageToServer(sendMessageToServer), _sendMessageToAllClient(sendMessageToAllClient)
 {
 }
 
@@ -18,36 +18,21 @@ RType::HandleShootSystem::~HandleShootSystem()
 
 void RType::HandleShootSystem::effects(std::vector<std::shared_ptr<RType::Entity>> entities) {
     for (const auto &entity: entities) {
-        if (verifyRequiredComponent(entity)) {
-            // auto playerGlobalBounds = entity->getComponent<SpriteComponent>()->getSprite()->getGlobalBounds();
-            if (entity->getComponent<ActionComponent>()->getActions(RType::SHOOTING) == true) {
+        if (verifyRequiredComponent(entity) && entity->getComponent<ActionComponent>()->getActions(RType::SHOOTING) == true) {
+            if (entity->getComponent<EntityTypeComponent>()->getEntityType() == E_PLAYER && _sendMessageToServer) {
                 entity->getComponent<RType::ActionComponent>()->setActions(RType::SHOOTING, false);
-
-                // std::shared_ptr<RType::Entity> shotEffect = _addEntity();
-                // shotEffect->pushComponent(std::make_shared<RType::EntityTypeComponent>(RType::E_BULLET_EFFECT));
-                // std::shared_ptr<RType::PositionComponent> position = shotEffect->pushComponent(std::make_shared<RType::PositionComponent>(playerGlobalBounds.width + playerGlobalBounds.left, (playerGlobalBounds.height / 2) + playerGlobalBounds.top));
-                // std::shared_ptr<RType::TextureComponent> texture = shotEffect->pushComponent(std::make_shared<RType::TextureComponent>("./ressources/effects/flash.png"));
-                // shotEffect->pushComponent(std::make_shared<RType::SpriteComponent>(texture->getTexture(), position->getPositions(), sf::Vector2f(2, 2), sf::IntRect(0, 0, 11, 19)));
-                // // shotEffect->pushComponent(std::make_shared<RType::DirectionPatternComponent>(RType::STRAIGHT_RIGHT));
-                // shotEffect->pushComponent(std::make_shared<RType::VelocityComponent>(7));
-                // shotEffect->pushComponent(std::make_shared<RType::ClockComponent>());
-
-                // std::shared_ptr<RType::Entity> shot = _addEntity();
-                // shot->pushComponent(std::make_shared<RType::EntityTypeComponent>(RType::E_BULLET));
-                // std::shared_ptr<RType::PositionComponent> position = shot->pushComponent(std::make_shared<RType::PositionComponent>(playerGlobalBounds.width + playerGlobalBounds.left, (playerGlobalBounds.height / 2) + playerGlobalBounds.top));
-                // std::shared_ptr<RType::TextureComponent> texture = shot->pushComponent(std::make_shared<RType::TextureComponent>("./ressources/shoot-spritesheet.png"));
-                // shot->pushComponent(std::make_shared<RType::SpriteComponent>(texture->getTexture(), position->getPositions(), sf::Vector2f(2, 2), sf::IntRect(0, 0, 19, 6)));
-                // shot->pushComponent(std::make_shared<RType::DirectionPatternComponent>(RType::STRAIGHT_RIGHT));
-                // shot->pushComponent(std::make_shared<RType::VelocityComponent>(7));
-                // shot->pushComponent(std::make_shared<RType::ClockComponent>());
-                // std::cout << "shoot" << std::endl;
                 _sendMessageToServer(Encoder::actionPlayer(true, false, false, false));
+            } else if (EntityTypeComponent::isMob(entity->getComponent<EntityTypeComponent>()->getEntityType()) && _sendMessageToAllClient
+                && entity->GET_CLOCK && entity->getComponent<ShootIntervalComponent>()
+                && entity->GET_SHOOTING_CLOCK.getElapsedTime().asSeconds() > entity->getComponent<ShootIntervalComponent>()->getShootInterval()) {
+                createEnnemyBullet(entity);
+                entity->GET_SHOOTING_CLOCK.restart();
             }
         }
-        if (entity->getComponent<RType::ClockComponent>() != nullptr && entity->getComponent<RType::EntityTypeComponent>()->getEntityType() == RType::E_BULLET_EFFECT) {
-            if (entity->getComponent<RType::ClockComponent>()->getClock(SHOOTING_EFFECT).getElapsedTime().asSeconds() >= 0.05) {
-                _deleteEntity(entity);
-            }
+        if (entity->getComponent<RType::ClockComponent>() != nullptr
+        && entity->getComponent<RType::EntityTypeComponent>()->getEntityType() == RType::E_BULLET_EFFECT
+        && entity->getComponent<RType::ClockComponent>()->getClock(SHOOTING_EFFECT).getElapsedTime().asSeconds() >= 0.05) {
+            _deleteEntity(entity);
         }
     }
 }
@@ -59,11 +44,28 @@ void RType::HandleShootSystem::effect(std::shared_ptr<RType::Entity> entity)
 
 bool RType::HandleShootSystem::verifyRequiredComponent(std::shared_ptr<RType::Entity> entity)
 {
-    if (entity->getComponent<RType::DirectionComponent>() == nullptr
-    || entity->getComponent<RType::ActionComponent>() == nullptr
-    || entity->getComponent<RType::SpriteComponent>() == nullptr)
-    {
+    if (entity->getComponent<RType::EntityTypeComponent>() == nullptr
+    || entity->getComponent<RType::ActionComponent>() == nullptr) {
         return false;
     }
     return (true);
+}
+void RType::HandleShootSystem::createEnnemyBullet(const std::shared_ptr<RType::Entity> &entity)
+{
+    std::shared_ptr<RType::Entity> bullet = _addEntity();
+
+    bullet->pushComponent(std::make_shared<RType::EntityTypeComponent>(RType::E_ENNEMY_BULLET));
+
+    float x = entity->getComponent<PositionComponent>()->getPositions().x;
+    float y = entity->getComponent<PositionComponent>()->getPositions().y + ((entity->getComponent<IntRectComponent>()->getIntRectHeight() * entity->getComponent<ScaleComponent>()->getScaleY()) / 2.);
+
+    std::shared_ptr<RType::PositionComponent> position = bullet->pushComponent(std::make_shared<RType::PositionComponent>(x, y));
+    bullet->pushComponent(std::make_shared<RType::DirectionPatternComponent>(RType::STRAIGHT_LEFT));
+    bullet->pushComponent(std::make_shared<RType::IntRectComponent>(0, 0, 19, 6));
+    bullet->pushComponent(std::make_shared<RType::ScaleComponent>(2.0, 2.0));
+    bullet->pushComponent(std::make_shared<RType::VelocityComponent>(21));
+    bullet->pushComponent(std::make_shared<RType::ClockComponent>());
+    bullet->pushComponent(std::make_shared<RType::MenuComponent>(GAME));
+
+    _sendMessageToAllClient(Encoder::newEntity(E_ENNEMY_BULLET, bullet->getId(), position->getPositionX(), position->getPositionY()));
 }
